+++ cmd/hidden_lake/service/pkg/app/init_database.go
@@ -5,12 +5,13 @@ import (
 
 	pkg_settings "github.com/number571/go-peer/cmd/hidden_lake/service/pkg/settings"
 	"github.com/number571/go-peer/pkg/errors"
+	"github.com/number571/go-peer/pkg/storage"
 	"github.com/number571/go-peer/pkg/storage/database"
 )
 
 func (p *sApp) initDatabase() error {
 	db, err := database.NewKeyValueDB(
-		database.NewSettings(&database.SSettings{
+		storage.NewSettings(&storage.SSettings{
 			FPath:      fmt.Sprintf("%s/%s", p.fPathTo, pkg_settings.CPathDB),
 			FHashing:   false,
 			FCipherKey: []byte("_"),