+++ cmd/hidden_lake/traffic/internal/database/database.go
@@ -7,6 +7,7 @@ import (
 	"github.com/number571/go-peer/pkg/crypto/hashing"
 	"github.com/number571/go-peer/pkg/encoding"
 	"github.com/number571/go-peer/pkg/errors"
+	"github.com/number571/go-peer/pkg/storage"
 	gp_database "github.com/number571/go-peer/pkg/storage/database"
 )
 
@@ -20,7 +21,7 @@ type sKeyValueDB struct {
 
 func NewKeyValueDB(pSett ISettings) (IKVDatabase, error) {
 	sqlDB, err := gp_database.NewKeyValueDB(
-		gp_database.NewSettings(&gp_database.SSettings{
+		storage.NewSettings(&storage.SSettings{
 			FPath:      pSett.GetPath(),
 			FHashing:   false,
 			FCipherKey: []byte("_"),