+++ examples/_modules/network/anonymity/echo/main.go
@@ -14,6 +14,7 @@ import (
 	"github.com/number571/go-peer/pkg/network/anonymity"
 	"github.com/number571/go-peer/pkg/network/anonymity/adapters"
 	"github.com/number571/go-peer/pkg/network/conn"
+	"github.com/number571/go-peer/pkg/storage"
 	"github.com/number571/go-peer/pkg/storage/database"
 )
 
@@ -76,7 +77,7 @@ func main() {
 
 func newNode(serviceAddress, dbPath string) anonymity.INode {
 	db, err := database.NewKeyValueDB(
-		database.NewSettings(&database.SSettings{
+		storage.NewSettings(&storage.SSettings{
 			FPath:      dbPath,
 			FHashing:   false,
 			FCipherKey: []byte("CIPHER"),