+++ pkg/storage/settings.go
@@ -7,14 +7,14 @@ var (
 type SSettings sSettings
 type sSettings struct {
 	FPath      string
-	FWorkSize  uint64
+	FHashing   bool
 	FCipherKey []byte
 }
 
 func NewSettings(pSett *SSettings) ISettings {
 	return (&sSettings{
 		FPath:      pSett.FPath,
-		FWorkSize:  pSett.FWorkSize,
+		FHashing:   pSett.FHashing,
 		FCipherKey: pSett.FCipherKey,
 	}).mustNotNull()
 }
@@ -22,9 +22,6 @@ func NewSettings(pSett *SSettings) ISettings {
 func (p *sSettings) mustNotNull() ISettings {
 	if p.FPath == "" {
 		panic(`p.FPath == ""`)
-	}
-	if p.FWorkSize == 0 {
-		panic(`p.FWorkSize == 0`)
 	}
 	if p.FCipherKey == nil {
 		panic(`p.FCipherKey == nil`)
@@ -36,8 +33,8 @@ func (p *sSettings) GetPath() string {
 	return p.FPath
 }
 
-func (p *sSettings) GetWorkSize() uint64 {
-	return p.FWorkSize
+func (p *sSettings) GetHashing() bool {
+	return p.FHashing
 }
 
 func (p *sSettings) GetCipherKey() []byte {